added volume integration tests for all the api version
